backend/requirements.txt
# FastAPI và dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
python-multipart==0.0.6

# AI/ML libraries
sentence-transformers==2.2.2
transformers==4.35.0
torch==2.1.0
faiss-cpu==1.7.4
numpy==1.24.3

# Document processing
PyPDF2==3.0.1
python-docx==1.1.0
python-docx2txt==0.8

# PDF processing with OCR
PyMuPDF==1.23.8
pdf2image==1.16.3
pytesseract==0.3.10
Pillow==10.1.0

# Database
sqlalchemy==2.0.23
alembic==1.12.1

# Utilities
aiofiles==23.2.1
python-dotenv==1.0.0
pydantic-settings==2.0.3

# Development
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2
